Binance-Price-Alert/requirements.txt
binance-connector
smtplib
